BigDecimal to double

+++ src/main/java/pl/serweron/serweronLib/api/economy/entity/IEconomyAccount.java
@@ -4,8 +4,6 @@ import pl.serweron.serweronLib.api.entity.IUser;
 import pl.serweron.serweronLib.economy.TransactionResult;
 import pl.serweron.serweronLib.utils.Response;
 
-import java.math.BigDecimal;
-
 /**
  * Represents an economy account tied to a user within the economy system.
  * Provides methods to access balance, perform transactions, and manage the account lifecycle.
@@ -15,14 +13,14 @@ public interface IEconomyAccount extends IUser {
     /**
      * Returns the current balance of the account.
      *
-     * @return the balance as a {@link BigDecimal}
+     * @return the balance as a {@link double}
      */
-    BigDecimal getBalance();
+    double getBalance();
 
     /**
      * Attempts to withdraw a specified amount from the account balance.
      *
-     * @param amount the amount to withdraw (using double for convenience, but BigDecimal recommended internally)
+     * @param amount the amount to withdraw (using double for convenience)
      * @return a {@link TransactionResult} indicating success, failure, or errors of the operation
      */
     TransactionResult withdraw(double amount);
@@ -41,7 +39,7 @@ public interface IEconomyAccount extends IUser {
      * @param amount the amount to check
      * @return true if the balance is greater than or equal to the amount, false otherwise
      */
-    boolean has(BigDecimal amount);
+    boolean has(double amount);
 
     /**
      * Deletes the economy account and performs necessary cleanup.
